+++ src/vec3.rs
@@ -1,24 +1,24 @@
 use rand::Rng;
 
 #[derive(Clone, Copy, Debug)]
-pub struct Vec3(pub f32, pub f32, pub f32);
+pub struct Vec3(pub f64, pub f64, pub f64);
 pub use Vec3 as Point3;
 pub use Vec3 as Color;
 
 impl Vec3 {
-  pub fn x(&self) -> f32 {
+  pub fn x(&self) -> f64 {
     self.0
   }
-  pub fn y(&self) -> f32 {
+  pub fn y(&self) -> f64 {
     self.1
   }
-  pub fn z(&self) -> f32 {
+  pub fn z(&self) -> f64 {
     self.2
   }
-  pub fn dot(&self, rhs: &Vec3) -> f32 {
+  pub fn dot(&self, rhs: &Vec3) -> f64 {
     (self.0 * rhs.0) + (self.1 * rhs.1) + (self.2 * rhs.2)
   }
-  pub fn cross(&self, rhs: &Vec3) -> f32 {
+  pub fn cross(&self, rhs: &Vec3) -> f64 {
     (self.1 * rhs.2 - self.2 * rhs.1)
       + (self.2 * rhs.0 - self.0 * rhs.2)
       + (self.0 * rhs.1 - self.1 * rhs.0)
@@ -26,13 +26,13 @@ impl Vec3 {
   pub fn unit(&self) -> Vec3 {
     *self / self.length()
   }
-  pub fn length(&self) -> f32 {
+  pub fn length(&self) -> f64 {
     self.length_squared().sqrt()
   }
-  pub fn length_squared(&self) -> f32 {
+  pub fn length_squared(&self) -> f64 {
     (self.0 * self.0) + (self.1 * self.1) + (self.2 * self.2)
   }
-  pub fn random_in_range(min: f32, max: f32) -> Vec3 {
+  pub fn random_in_range(min: f64, max: f64) -> Vec3 {
     let mut rng = rand::thread_rng();
     Vec3(
       rng.gen_range(min..max),
@@ -80,23 +80,23 @@ impl std::ops::Mul for Vec3 {
   }
 }
 
-impl std::ops::Mul<f32> for Vec3 {
+impl std::ops::Mul<f64> for Vec3 {
   type Output = Vec3;
-  fn mul(self, rhs: f32) -> Vec3 {
+  fn mul(self, rhs: f64) -> Vec3 {
     Vec3(self.0 * rhs, self.1 * rhs, self.2 * rhs)
   }
 }
 
-impl std::ops::Mul<Vec3> for f32 {
+impl std::ops::Mul<Vec3> for f64 {
   type Output = Vec3;
   fn mul(self, rhs: Vec3) -> Vec3 {
     Vec3(self * rhs.0, self * rhs.1, self * rhs.2)
   }
 }
 
-impl std::ops::Div<f32> for Vec3 {
+impl std::ops::Div<f64> for Vec3 {
   type Output = Vec3;
-  fn div(self, rhs: f32) -> Vec3 {
+  fn div(self, rhs: f64) -> Vec3 {
     Vec3(self.0 / rhs, self.1 / rhs, self.2 / rhs)
   }
 }
